refactor: move init to new subdirectory
