+++ packages/myst-cli/src/init/init.ts
@@ -2,16 +2,17 @@ import fs from 'node:fs';
 import path from 'node:path';
 import yaml from 'js-yaml';
 import { v4 as uuid } from 'uuid';
+import inquirer from 'inquirer';
+import chalk from 'chalk';
 import { defaultConfigFile, loadConfig, writeConfigs } from '../config.js';
 import { loadProjectFromDisk } from '../project/load.js';
 import { selectors } from '../store/index.js';
 import type { ISession } from '../session/types.js';
-import inquirer from 'inquirer';
-import chalk from 'chalk';
-import { startServer } from './site/start.js';
+import { startServer } from '../build/site/start.js';
 import { githubCurvenoteAction, githubPagesAction } from './gh-actions/index.js';
-import { getGithubUrl } from './utils/github.js';
-import { checkFolderIsGit } from './utils/git.js';
+import { getGithubUrl } from '../utils/github.js';
+import { checkFolderIsGit } from '../utils/git.js';
+
 const VERSION_CONFIG = '# See docs at: https://mystmd.org/guide/frontmatter\nversion: 1\n';
 
 function createProjectConfig({ github }: { github?: string } = {}) {
