+++ packages/myst-cli/src/utils/index.ts
@@ -12,6 +12,7 @@ export * from './resolveExtension.js';
 export * from './shouldIgnoreFile.js';
 export * from './toc.js';
 export * from './uniqueArray.js';
+export * from './github.js';
 
 export * as imagemagick from './imagemagick.js';
 export * as inkscape from './inkscape.js';
